Web scraping using python
The project uses the beautiful soup python library to scrape the timejobs website to get information about the latest python developer jobs. The program provides the company name, extra required skills, and website link to apply for that post. Additionally, it also omits the jobs that required the skill set unknown to the person. Lastly, the data scraped is stored into text files in the posts directory.